chore(homepage): move all apis from admin to content-manager
