--- packages/core/admin/server/src/routes/homepage.ts
@@ -1,15 +1,0 @@
-import type { Plugin } from '@strapi/types';
-
-const info = { pluginName: 'admin', type: 'admin' };
-
-export default [
-  {
-    method: 'GET',
-    info,
-    path: '/homepage/recent-documents',
-    handler: 'homepage.getRecentDocuments',
-    config: {
-      policies: ['admin::isAuthenticatedAdmin'],
-    },
-  },
-] satisfies Plugin.LoadedPlugin['routes'][string]['routes'];
