+++ packages/core/admin/server/src/routes/index.ts
@@ -7,7 +7,6 @@ import webhooks from './webhooks';
 import apiTokens from './api-tokens';
 import contentApi from './content-api';
 import transfer from './transfer';
-import homepage from './homepage';
 
 const routes = {
   admin: {
@@ -22,7 +21,6 @@ const routes = {
       ...apiTokens,
       ...contentApi,
       ...transfer,
-      ...homepage,
     ],
   },
 };
