+++ packages/core/admin/server/src/services/index.ts
@@ -13,7 +13,6 @@ import * as action from './action';
 import * as apiToken from './api-token';
 import * as transfer from './transfer';
 import * as projectSettings from './project-settings';
-import { createHomepageService } from './homepage';
 
 // TODO: TS - Export services one by one as this export is cjs
 export default {
@@ -31,5 +30,4 @@ export default {
   'api-token': apiToken,
   transfer,
   'project-settings': projectSettings,
-  homepage: createHomepageService,
 };
