+++ packages/core/admin/server/src/utils/index.d.ts
@@ -9,7 +9,6 @@ import * as token from '../services/token';
 import * as apiToken from '../services/api-token';
 import * as projectSettings from '../services/project-settings';
 import * as transfer from '../services/transfer';
-import { createHomepageService } from '../services/homepage';
 
 type S = {
   role: typeof role;
@@ -23,7 +22,6 @@ type S = {
   'api-token': typeof apiToken;
   'project-settings': typeof projectSettings;
   transfer: typeof transfer;
-  homepage: ReturnType<typeof createHomepageService>;
 };
 
 type Resolve<T> = T extends (...args: unknown[]) => unknown ? T : { [K in keyof T]: T[K] };
