+++ packages/core/content-manager/server/src/homepage/index.ts
@@ -1,0 +1,9 @@
+import { routes } from './routes';
+import { controllers } from './controllers';
+import { services } from './services';
+
+export default {
+  routes,
+  controllers,
+  services,
+};
