+++ packages/core/content-manager/server/src/homepage/services/index.ts
@@ -1,0 +1,7 @@
+import type { Plugin } from '@strapi/types';
+
+import { createHomepageService } from './homepage';
+
+export const services = {
+  homepage: createHomepageService,
+} satisfies Plugin.LoadedPlugin['services'];
