+++ packages/core/content-manager/server/src/routes/index.ts
@@ -1,9 +1,11 @@
 import admin from './admin';
 import history from '../history';
 import preview from '../preview';
+import homepage from '../homepage';
 
 export default {
   admin,
   ...(history.routes ? history.routes : {}),
   ...(preview.routes ? preview.routes : {}),
+  ...homepage.routes,
 };
