+++ packages/core/content-manager/server/src/services/index.ts
@@ -9,6 +9,7 @@ import populateBuilder from './populate-builder';
 import uid from './uid';
 import history from '../history';
 import preview from '../preview';
+import homepage from '../homepage';
 import documentMetadata from './document-metadata';
 import documentManager from './document-manager';
 
@@ -26,4 +27,5 @@ export default {
   uid,
   ...(history.services ? history.services : {}),
   ...(preview.services ? preview.services : {}),
+  ...homepage.services,
 };
